refactor: constants
